cortex: Streamline and silence health check
The cortex container healthcheck was quite "loud" in the sense that each
check produced about 20 lines of Cortex log output regarding missing
authentication. We silence that by using the Peekaboo API token for
these checks. That way they are authenticated and produce status code
200 instead of 401. That also means, we implicitly check correct setup
because the API token must exist and be correct for the container to
become healthy. The need for initial setup can still be detected by
status code 520 being returned by Cortex.

While at it, we streamline the check using grep and make it more
readable and reduce the need for escaping by switching to a yaml
multiline block.