rspamd: Make peekaboo plugin more accessible
Since we add the peekaboo rspamd plugin as a whole anyway, there's no
use in hiding it in the integration patch. So we move it out into a file
for easier review, comparison and update. The integration patch
continues to contain changes to rspamd-supplied files needed to make the
plugin work. The integration patch file is put into /root inside the
image to not pollute the configuration directory and because it's small
enough not to warrant a multi-stage build to keep it out of the image.