ci/cd: Add workflow to build and publish container images
Add a workflow for building and publishing container images based on the
example Github Actions template but modify it to accept the image name
and build context directory as parameters so we can potentially build
multiple images with it. Add another workflow that uses the reusable one
to build and publish our images. In the case of the PeekabooAV image we
always supply the latest devel version from its repo for now.