ci/cd: Add workflow to build and publish container images
Add a workflow for building and publishing container images based on the
example Github Actions template but turn it into a composite action and
modify it to accept the image name and build context directory as
parameters so we can build multiple images with it. Add another workflow
that uses the composite action to build and publish our images using a
matrix build.