pipeline, ansible: Switch to drop files for config
Switch configuration mechanics in the Ansible-based installer and
docker-compose-based pipeline to drop-files from scVENUS/PeekabooAV#212.

The main config files are supplied as copies of the distributed sample
configs. This is so that users can refer to the comments therein for
reference (although somewhat removed from access in the container
use-case). This allows us to reduce the respective installer-supplied
config snippets to the bare minimum of what we want to override. It
should remove much of the need to keep copies in sync during
development. It also allows users to add their own drop files which
survive re-runs of the installer or other forms of updates.

Both the installer and container entrypoint do their own configuration
through drop-files themselves. These are named 05-installer.conf and
05-entrypoint.conf respectively. This allows to add a few lower- and
quite some higher-priority drop files before and after in
lexicographical sorting order. In the case of the pipeline we also add a
direct volume mount of a ruleset.conf.d/10-compose.conf override drop
file that replaces parts of the sample ruleset relevant for that
use-case (the entrypoint does not support creation of a ruleset config
drop file as of now for complexity reasons).