rspamd: Generalise container image
Baking config into the container image is not ideal. Mounting config via
volumes from the outside is complicated by the elaborate include file
structure of rspamd. So we jump straight to 4000% and extend the
entrypoint to allow setting arbitrary config options via environment
variables. The details are explained in comments in the entrypoint. A
simple debug mechanism is added based on the presence of a special
variable RSPAMD_DEBUG_CONFIG which prints out all generated config files
before starting rspamd. The configuration snippets formerly baked into
the image are moved into the compose file.