rspamd, postfix: Add no cache flag to apk
The --no-cache flag to apk actually saves a couple of more megabytes in
image size. So we add it where appropriate.

In the case of postfix we remove mailutils because only swaks is really
used. Also, we enable the testing repository only for installing swaks
so that we get all other packages from the stable repositories.